Move test dataset to inst/extdata so it can be conveniently loaded.
The normal data loading from data in 'data' assumes that CSV files use a semi-colon delimiter, and so our data file does not load properly.  In addition, it is difficult to get the data dir to load in both non-installed and installed packages in a consistent way. By relocating our test dataset to inst/extdata, both test() and test_package() functions work.