Revert back to tests in inst/tests.
Although the new convention is for tests to be in tests/testthat, and to use test_check() rather than test_package(), I could not get this to work in all scenarios. Reverting to their location in inst/tests which allows tests to run under devtools, under R CMD check, and on an installed package using test_package().